fix: SSH reverse tunnel detection, clippy cleanup
- Shield: fixed ssh -R detection (lowercase context, was duplicate -l)
- Clippy: auto-fixed 15+ style warnings (redundant borrows, contains vs any)
- Removed unused mut, simplified boolean expressions
- All tests green, 0 warnings, 0 clippy errors

+++ crates/agent/src/sandbox.rs
@@ -403,7 +403,7 @@ fn create_dev_node(
         let ret = libc::mknod(path_cstr.as_ptr(), mode, dev);
         if ret != 0 {
             let errno = std::io::Error::last_os_error().raw_os_error().unwrap_or(0);
-            if errno != libc::EEXIST as i32 {
+            if errno != libc::EEXIST {
                 log::debug!("mknod for {} failed (errno={}): may need root", name, errno);
             }
         }
